[Pal/Linux-SGX] Remove unnecessary check on exit_target
exit_target is only used as an argument to EEXIT to tell the processor
where to continue execution after leaving enclave mode. As documented in
the Intel SDM: If it points to enclave memory it will be restored but
the access will happen after leaving enclave mode so the usual enclave
protections are enforced. Therefore the software check is unnecessary.

Also size == 0 didn't make much sense.