[Pal/Linux-SGX] Save/restore callee-saved regs on sgx_{ecall,raise}
The EEXIT instruction does not restore registers of untrusted part of
the program. Thus, sgx_ecall() and sgx_raise() functions must save
callee-saved regs before EENTER and restore them after EEXIT. This
fixes the bug of a build with DEBUG=0 when the untrusted part of the
program uses RBP and other regs (in contrast to undefined DEBUG when
these registers are optimized away from the untrusted part).