[Pal/lib] Build in-Graphene mbedTLS crypto lib with minimal config
Graphene now moved to mbedTLS version 2.21.0. This new version
decouples the crypto functionality in a separate lib (libmbedcrypto).
This separate lib has its C code and headers under mbedtls/crypto
subdirectory. Previously, we forgot to copy our in-Graphene minimal
config `config.h` into this new subdirectory. This meant that the
crypto lib was built with a default and bloated configuration, not
needed for our Graphene purposes. This commit fixes this.