[Pal/Linux-SGX] tools: add Secret Provisioning utility libraries
Secret Provisioning provides a simple interface for enclavized
applications to connect to a trusted secret-provisioning service and
securely pass the secrets inside the SGX enclave. The application acts
as a client and the secret-provisioning service as a server. They
establish a TLS session via mutual authentication, where the service
sends a classical X.509 certificate and the application sends an
RA-TLS X.509 certificate. After the service verifies the RA-TLS
certificate and its SGX-related information, it provisions secrets
to the application. The TLS session may continue to be used by the
application and the server after the initial provisioning. Secret
Provisioning feature relies heavily on RA-TLS.

Secret Provisioning is shipped as three libraries:

- secret_prov_attest.so: creates RA-TLS X.509 certificate with SGX
  quote embedded and sends it to the secret-provisioning service as
  part of mutual attestation of TLS session. Linked into enclavized
  applications. Not thread-safe.
- secret_prov_verify_epid.so: establishes a TLS session with mutual
  attestation, verifies received RA-TLS certificate, and sends secrets
  back to the enclavized application. Linked into secret-provisioning
  service. Verifies RA-TLS certificate using EPID-based flows.
  Multi-threaded and thread-safe (new thread for each client).
- secret_prov_verify_dcap.so: same as secret_prov_verify_epid.so but
  verifies RA-TLS certificate using ECDSA/DCAP-based flows.
  Multi-threaded and thread-safe (new thread for each client).

This commit also adds two Secret Provisioning examples, with a minimal
enclavized app, a more complicated enclavized app, and a single
secret-provisioning service. These examples are added to Jenkins.